Fixed float comparison issue introduced in 1.9.4 PRB 1.  (made nit, br and f$P in _iter_params into unsigned ints, and adjusted other affected code).  Thanks to flounderhead for finding this and suggesting a fix.

+++ trunk/OpenHoldem/CIteratorThread.h
@@ -8,9 +8,9 @@ extern CRITICAL_SECTION	cs_iterator;
 
 extern struct SIterParams
 {
-	double			nit;
-	double			f$p;
-	double			br;
+	unsigned int	nit;
+	unsigned int	f$p;
+	unsigned int	br;
 	unsigned int	pcard[2];
 	unsigned int	ccard[5];
 } _iter_params;
